requirements.txt
Flask==3.0.3
openai==1.42.0
python-docx==1.1.2
PyPDF2==3.0.1
Werkzeug==3.0.4
reportlab==4.2.0
httpx==0.27.2
